Restructure package from sarixfourier to sarix
- Rename package from sarixfourier to sarix (v0.1.0 -> v0.2.0)
- Move module from src/sarixfourier/ to src/sarix/
- Update tests to match new module structure
- Add CHANGELOG.md to track version history

This change simplifies the package name while maintaining all functionality.
