+++ pyproject.toml
@@ -3,8 +3,8 @@ requires = ["setuptools>=45", "wheel", "setuptools-scm[toml]>=6.2"]
 build-backend = "setuptools.build_meta"
 
 [project]
-name = "sarixfourier"
-version = "0.1.0"
+name = "sarix"
+version = "0.2.0"
 description = "Seasonal AR, Integrated models with eXogenous predictors, with Fourier seasonality support"
 readme = "README.md"
 requires-python = ">=3.11"
