+++ src/sarix/__init__.py
@@ -1,0 +1,4 @@
+from sarix.sarix import SARIX, diff, inv_diff
+
+__all__ = ['SARIX', 'diff', 'inv_diff']
+__version__ = '0.2.0'
