--- src/sarixfourier/__init__.py
@@ -1,4 +1,0 @@
-from sarixfourier.sarix_fourier import SARIX, diff, inv_diff
-
-__all__ = ['SARIX', 'diff', 'inv_diff']
-__version__ = '0.1.0'
